// src/plugins/doordash/generated-admin-types.ts
export type InputMaybe<T> = T;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: Date; output: string; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type DoordashConfigFilterParameter = {
  _and?: InputMaybe<Array<DoordashConfigFilterParameter>>;
  _or?: InputMaybe<Array<DoordashConfigFilterParameter>>;
  apiEndpoint?: InputMaybe<StringOperators>;
  createdAt?: InputMaybe<DateOperators>;
  developerId?: InputMaybe<StringOperators>;
  enabled?: InputMaybe<BooleanOperators>;
  id?: InputMaybe<IdOperators>;
  keyId?: InputMaybe<StringOperators>;
  sandbox?: InputMaybe<BooleanOperators>;
  signingSecret?: InputMaybe<StringOperators>;
  state?: InputMaybe<StringOperators>;
  status?: InputMaybe<StringOperators>;
  updatedAt?: InputMaybe<DateOperators>;
};

export type DoordashConfigList = PaginatedList & {
  __typename?: 'DoordashConfigList';
  items: Array<DoordashConfig>;
  totalItems: Scalars['Int']['output'];
};

export type DoordashConfigListOptions = {
  /** Allows the results to be filtered */
  filter?: InputMaybe<DoordashConfigFilterParameter>;
  /** Specifies whether multiple top-level "filter" fields should be combined with a logical AND or OR operation. Defaults to AND. */
  filterOperator?: InputMaybe<LogicalOperator>;
  /** Skips the first n results, for use in pagination */
  skip?: InputMaybe<Scalars['Int']['input']>;
  /** Specifies which properties to sort the results by */
  sort?: InputMaybe<DoordashConfigSortParameter>;
  /** Takes n results, for use in pagination */
  take?: InputMaybe<Scalars['Int']['input']>;
};

export type DoordashConfigSortParameter = {
  apiEndpoint?: InputMaybe<SortOrder>;
  createdAt?: InputMaybe<SortOrder>;
  developerId?: InputMaybe<SortOrder>;
  id?: InputMaybe<SortOrder>;
  keyId?: InputMaybe<SortOrder>;
  signingSecret?: InputMaybe<SortOrder>;
  state?: InputMaybe<SortOrder>;
  status?: InputMaybe<SortOrder>;
  updatedAt?: InputMaybe<SortOrder>;
};
// ...
export enum LogicalOperator {
  AND = 'AND',
  OR = 'OR'
}
// ...
export type MutationSubmitDoordashConfigArgs = {
  input: SubmitDoordashConfigInput;
};
// ...
export type MutationUpsertDoordashConfigArgs = {
  input: UpdateDoordashConfigInput;
};
// ...
export type QueryDoordashConfigArgs = {
  id: Scalars['ID']['input'];
};


export type QueryDoordashConfigsArgs = {
  options?: InputMaybe<DoordashConfigListOptions>;
};


export type QueryEligibleShippingMethodsForDraftOrderArgs = {
  orderId: Scalars['ID']['input'];
};


export type QueryFacetArgs = {
  id: Scalars['ID']['input'];
};


export type QueryFacetValuesArgs = {
  options?: InputMaybe<FacetValueListOptions>;
};


export type QueryFacetsArgs = {
  options?: InputMaybe<FacetListOptions>;
};


export type QueryIsDoordashConfigValidArgs = {
  input: SubmitDoordashConfigInput;
};
// ...
export type SubmitDoordashConfigInput = {
  apiEndpoint: Scalars['String']['input'];
  developerId: Scalars['String']['input'];
  enabled: Scalars['Boolean']['input'];
  id?: InputMaybe<Scalars['ID']['input']>;
  keyId: Scalars['String']['input'];
  sandbox: Scalars['Boolean']['input'];
  signingSecret: Scalars['String']['input'];
};
// ...
export type UpdateDoordashConfigInput = {
  apiEndpoint: Scalars['String']['input'];
  developerId: Scalars['String']['input'];
  enabled: Scalars['Boolean']['input'];
  id: Scalars['ID']['input'];
  keyId: Scalars['String']['input'];
  sandbox: Scalars['Boolean']['input'];
  signingSecret: Scalars['String']['input'];
};